Make --tree mode work without --dirs.
Turns out that the NLTK trees needed to be post-processed to work properly
with zss's notion of how trees work, and that post-processing was only done in
the `aggregate_tree` method called when the --dirs option is active.

This factors the tree munging out into `read_tree` so that it happens in all
cases.